Reworked generator chaining.
