+++ pyEDAA/OutputFilter/Xilinx/__init__.py
@@ -29,379 +29,31 @@
 # ==================================================================================================================== #
 #
 """Basic classes for outputs from AMD/Xilinx Vivado."""
-from datetime import datetime
-from enum     import Flag
 from pathlib  import Path
-from re       import compile as re_compile, Pattern
-from typing import ClassVar, Self, Optional as Nullable, Dict, Type, Callable, List, Generator, Union, Tuple
+from typing import Optional as Nullable, Dict, List, Generator, Union, Type
 
 from pyTooling.Decorators  import export, readonly
-from pyTooling.MetaClasses import ExtendedType, abstractmethod
+from pyTooling.MetaClasses import ExtendedType
 from pyTooling.Stopwatch   import Stopwatch
-from pyTooling.TerminalUI  import TerminalApplication
-from pyTooling.Versioning  import YearReleaseVersion
 
-from pyEDAA.OutputFilter   import OutputFilterException
+from pyEDAA.OutputFilter.Xilinx.Common import LineKind, Line, InfoMessage, WarningMessage, CriticalWarningMessage, \
+	ErrorMessage, VivadoMessage, VivadoInfoMessage, VivadoIrregularInfoMessage, VivadoWarningMessage, \
+	VivadoIrregularWarningMessage, VivadoCriticalWarningMessage, VivadoErrorMessage, VHDLReportMessage, VivadoTclCommand, \
+	TclCommand
+from pyEDAA.OutputFilter.Xilinx.Commands import Command, SynthesizeDesign
+from pyEDAA.OutputFilter.Xilinx.Common2 import Preamble
+from pyEDAA.OutputFilter.Xilinx.Exception import ProcessorException, ClassificationException
 
 
-@export
-class ProcessorException(OutputFilterException):
-	pass
-
-
-@export
-class ClassificationException(ProcessorException):
-	_lineNumber: int
-	_rawMessage: str
-
-	def __init__(self, errorMessage: str, lineNumber: int, rawMessageLine: str):
-		super().__init__(errorMessage)
-
-		self._lineNumber = lineNumber
-		self._rawMessage = rawMessageLine
+ProcessedLine = Union[Line, ProcessorException]
 
 
 @export
-class ParserStateException(ProcessorException):
-	pass
-
-
-@export
-class LineKind(Flag):
-	Unprocessed =                0
-	ProcessorError =         2** 0
-	Empty =                  2** 1
-	Delimiter =              2** 2
-
-	Verbose =                2**10
-	Normal =                 2**11
-	Info =                   2**12
-	Warning =                2**13
-	CriticalWarning =        2**14
-	Error =                  2**15
-	Fatal =                  2**16
-
-	Start =                  2**20
-	End =                    2**21
-	Header =                 2**22
-	Content =                2**23
-	Footer =                 2**24
-
-	Last =                   2**29
-
-	Message =                2**30
-	InfoMessage =            Message | Info
-	WarningMessage =         Message | Warning
-	CriticalWarningMessage = Message | CriticalWarning
-	ErrorMessage =           Message | Error
-
-	Section =                2**31
-	SectionDelimiter =       Section | Delimiter
-	SectionStart =           Section | Start
-	SectionEnd =             Section | End
-
-	SubSection =             2**32
-	SubSectionDelimiter =    SubSection | Delimiter
-	SubSectionStart =        SubSection | Start
-	SubSectionEnd =          SubSection | End
-
-	Paragraph =              2**33
-	ParagraphHeadline =      Paragraph | Header
-
-	Table =                  2**34
-	TableFrame =             Table | Delimiter
-	TableHeader =            Table | Header
-	TableRow =               Table | Content
-	TableFooter =            Table | Footer
-
-	Command =                2**35
-
-
-@export
-class Line(metaclass=ExtendedType, slots=True):
-	"""
-	This class represents any line in a log file.
-	"""
-	_lineNumber:    int
-	_kind:          LineKind
-	_message:       str
-
-	def __init__(self, lineNumber: int, kind: LineKind, message: str) -> None:
-		self._lineNumber = lineNumber
-		self._kind = kind
-		self._message = message
-
-	@readonly
-	def LineNumber(self) -> int:
-		return self._lineNumber
-
-	@readonly
-	def Kind(self) -> LineKind:
-		return self._kind
-
-	@readonly
-	def Message(self) -> str:
-		return self._message
-
-	def __str__(self) -> str:
-		return self._message
-
-
-@export
-class InfoMessage(metaclass=ExtendedType, mixin=True):
-	pass
-
-
-@export
-class WarningMessage(metaclass=ExtendedType, mixin=True):
-	pass
-
-
-@export
-class CriticalWarningMessage(metaclass=ExtendedType, mixin=True):
-	pass
-
-
-@export
-class ErrorMessage(metaclass=ExtendedType, mixin=True):
-	pass
-
-
-@export
-class VivadoMessage(Line):
-	"""
-	This class represents an AMD/Xilinx Vivado message.
-
-	The usual message format is:
-
-	.. code-block:: text
-
-	   INFO: [Synth 8-7079] Multithreading enabled for synth_design using a maximum of 2 processes.
-	   WARNING: [Synth 8-3332] Sequential element (gen[0].Sync/FF2) is unused and will be removed from module sync_Bits_Xilinx.
-
-	The following message severities are defined:
-
-	* ``INFO``
-	* ``WARNING``
-	* ``CRITICAL WARNING``
-	* ``ERROR``
-
-	.. seealso::
-
-	   :class:`VivadoInfoMessage`
-	     Representing a Vivado info message.
-
-	   :class:`VivadoWarningMessage`
-	     Representing a Vivado warning message.
-
-	   :class:`VivadoCriticalWarningMessage`
-	     Representing a Vivado critical warning message.
-
-	   :class:`VivadoErrorMessage`
-	     Representing a Vivado error message.
-	"""
-	# _MESSAGE_KIND:  ClassVar[str]
-	# _REGEXP:        ClassVar[Pattern]
-
-	_toolID:        int
-	_toolName:      str
-	_messageKindID: int
-
-	def __init__(self, lineNumber: int, kind: LineKind, tool: str, toolID: int, messageKindID: int, message: str) -> None:
-		super().__init__(lineNumber, kind, message)
-		self._toolID = toolID
-		self._toolName = tool
-		self._messageKindID = messageKindID
-
-	@readonly
-	def ToolName(self) -> str:
-		return self._toolName
-
-	@readonly
-	def ToolID(self) -> int:
-		return self._toolID
-
-	@readonly
-	def MessageKindID(self) -> int:
-		return self._messageKindID
-
-	@classmethod
-	def Parse(cls, lineNumber: int, kind: LineKind, rawMessage: str) -> Nullable[Self]:
-		if (match := cls._REGEXP.match(rawMessage)) is not None:
-			return cls(lineNumber, kind, match[1], int(match[2]), int(match[3]), match[4])
-
-		return None
-
-	def __str__(self) -> str:
-		return f"{self._MESSAGE_KIND}: [{self._toolName} {self._toolID}-{self._messageKindID}] {self._message}"
-
-
-@export
-class VivadoInfoMessage(VivadoMessage, InfoMessage):
-	"""
-	This class represents an AMD/Xilinx Vivado info message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "INFO"
-	_REGEXP:       ClassVar[Pattern] = re_compile(r"""INFO: \[(\w+) (\d+)-(\d+)\] (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		return super().Parse(lineNumber, LineKind.InfoMessage, rawMessage)
-
-
-@export
-class VivadoIrregularInfoMessage(VivadoMessage, InfoMessage):
-	"""
-	This class represents an irregular AMD/Xilinx Vivado info message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "INFO"
-	_REGEXP:      ClassVar[Pattern] = re_compile(r"""INFO: \[(\w+)-(\d+)\] (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		if (match := cls._REGEXP.match(rawMessage)) is not None:
-			return cls(lineNumber, LineKind.InfoMessage, match[1], None, int(match[2]), match[3])
-
-		return None
-
-	def __str__(self) -> str:
-		return f"{self._MESSAGE_KIND}: [{self._toolName}-{self._messageKindID}] {self._message}"
-
-
-@export
-class VivadoWarningMessage(VivadoMessage, WarningMessage):
-	"""
-	This class represents an AMD/Xilinx Vivado warning message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "WARNING"
-	_REGEXP:       ClassVar[Pattern] = re_compile(r"""WARNING: \[(\w+) (\d+)-(\d+)\] (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		return super().Parse(lineNumber, LineKind.WarningMessage, rawMessage)
-
-
-@export
-class VivadoIrregularWarningMessage(VivadoMessage, WarningMessage):
-	"""
-	This class represents an AMD/Xilinx Vivado warning message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "WARNING"
-	_REGEXP:       ClassVar[Pattern] = re_compile(r"""WARNING: (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		if (match := cls._REGEXP.match(rawMessage)) is not None:
-			return cls(lineNumber, LineKind.WarningMessage, None, None, None, match[1])
-
-		return None
-
-	def __str__(self) -> str:
-		return f"{self._MESSAGE_KIND}: {self._message}"
-
-
-@export
-class VivadoCriticalWarningMessage(VivadoMessage, CriticalWarningMessage):
-	"""
-	This class represents an AMD/Xilinx Vivado critical warning message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "CRITICAL WARNING"
-	_REGEXP:       ClassVar[Pattern] = re_compile(r"""CRITICAL WARNING: \[(\w+) (\d+)-(\d+)\] (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		return super().Parse(lineNumber, LineKind.CriticalWarningMessage, rawMessage)
-
-
-@export
-class VivadoErrorMessage(VivadoMessage, ErrorMessage):
-	"""
-	This class represents an AMD/Xilinx Vivado error message.
-	"""
-
-	_MESSAGE_KIND: ClassVar[str] =     "ERROR"
-	_REGEXP:       ClassVar[Pattern] = re_compile(r"""ERROR: \[(\w+) (\d+)-(\d+)\] (.*)""")
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		return super().Parse(lineNumber, LineKind.ErrorMessage, rawMessage)
-
-
-@export
-class VHDLReportMessage(VivadoInfoMessage):
-	_REGEXP: ClassVar[Pattern ] = re_compile(r"""RTL report: "(.*)" \[(.*):(\d+)\]""")
-
-	_reportMessage:    str
-	_sourceFile:       Path
-	_sourceLineNumber: int
-
-	def __init__(self, lineNumber: int, tool: str, toolID: int, messageKindID: int, rawMessage: str, reportMessage: str, sourceFile: Path, sourceLineNumber: int):
-		super().__init__(lineNumber, LineKind.InfoMessage, tool, toolID, messageKindID, rawMessage)
-
-		self._reportMessage = reportMessage
-		self._sourceFile = sourceFile
-		self._sourceLineNumber = sourceLineNumber
-
-	@classmethod
-	def Convert(cls, line: VivadoInfoMessage) -> Nullable[Self]:
-		if (match := cls._REGEXP.match(line._message)) is not None:
-			return cls(line._lineNumber, line._toolName, line._toolID, line._messageKindID, line._message, match[1], Path(match[2]), int(match[3]))
-
-		return None
-
-
-@export
-class VHDLAssertionMessage(VHDLReportMessage):
-	_REGEXP: ClassVar[Pattern ] = re_compile(r"""RTL assertion: "(.*)" \[(.*):(\d+)\]""")
-
-
-@export
-class VivadoTclCommand(Line):
-	_PREFIX: ClassVar[str] = "Command:"
-
-	_command: str
-	_args:    Tuple[str, ...]
-
-	def __init__(self, lineNumber: int, command: str, arguments: Tuple[str, ...], tclCommand: str) -> None:
-		super().__init__(lineNumber, LineKind.Command, tclCommand)
-
-		self._command = command
-		self._args = arguments
-
-	@classmethod
-	def Parse(cls, lineNumber: int, rawMessage: str) -> Nullable[Self]:
-		command = rawMessage[len(cls._PREFIX) + 1:]
-		args = command.split()
-
-		return cls(lineNumber, args[0], tuple(args[1:]), command)
-
-	def __str__(self) -> str:
-		return f"{self._PREFIX} {self._command} {' '.join(self._args)}"
-
-
-@export
-class ProcessingState(Flag):
-	Processed =        1
-	Skipped =          2
-	EmptyLine =        4
-	CommentLine =      8
-	DelimiterLine =   16
-	TableLine =       32
-	TableHeader =     64
-	Reprocess =      512
-	Last =          1024
-
-
-@export
-class BaseProcessor(metaclass=ExtendedType, slots=True):
-	# _parsers:  Dict[Type["Parser"], "Parsers"]
-	# _state:    Callable[[int, str], bool]
-	_duration: float
+class Processor(metaclass=ExtendedType, slots=True):
+	_duration:                float
+
+	_preamble:                Preamble
+	_commands:                Dict[Type[Command], Command]
 
 	_infoMessages:            List[VivadoInfoMessage]
 	_warningMessages:         List[VivadoWarningMessage]
@@ -412,9 +64,10 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 	_messagesByID:            Dict[int, Dict[int, List[VivadoMessage]]]
 
 	def __init__(self):
-		# self._parsers =  {}
-		# self._state =    None
-		self._duration = 0.0
+		self._duration =                0.0
+
+		self._preamble =                None
+		self._commands =                {}
 
 		self._infoMessages =            []
 		self._warningMessages =         []
@@ -456,6 +109,7 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 	def ErrorMessages(self) -> List[VivadoErrorMessage]:
 		return self._errorMessages
 
+	# TODO: Synthesis specific !!
 	@readonly
 	def VHDLReportMessages(self) -> List[VHDLReportMessage]:
 		if 8 in self._messagesByID:
@@ -472,8 +126,8 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 
 		return []
 
-	# def __getitem__(self, item: Type["Parser"]) -> "Parsers":
-	# 	return self._parsers[item]
+	def __getitem__(self, item: Type[Command]) -> Command:
+		return self._commands[item]
 
 	def _AddMessageByID(self, message: VivadoMessage) -> None:
 		if message._toolID in self._messagesByID:
@@ -487,15 +141,14 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 			self._toolNames[message._toolName] = message._toolID
 			self._messagesByID[message._toolID] = {message._messageKindID: [message]}
 
-	def LineClassification(self, documentSlicer: Generator[Union[Line, ProcessorException], Line, None]) -> Generator[Union[Line, ProcessorException], str, None]:
-		# Initialize generator
-		next(documentSlicer)
+	def LineClassification(self) -> Generator[ProcessedLine, str, None]:
+		# Instantiate and initialize CommandFinder
+		next(cmdFinder := self.CommandFinder())
 
 		# wait for first line
 		rawMessageLine = yield
 		lineNumber = 0
 		_errorMessage = "Unknown processing error."
-		errorMessage = _errorMessage
 
 		while rawMessageLine is not None:
 			lineNumber += 1
@@ -531,7 +184,7 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 			if line is None:
 				line = Line(lineNumber, LineKind.ProcessorError, rawMessageLine)
 
-			line = documentSlicer.send(line)
+			line = cmdFinder.send(line)
 
 			if isinstance(line, VivadoMessage):
 				self._AddMessageByID(line)
@@ -549,75 +202,46 @@ class BaseProcessor(metaclass=ExtendedType, slots=True):
 
 			rawMessageLine = yield line
 
+	def CommandFinder(self) -> Generator[Line, Line, None]:
+		self._preamble = Preamble(self)
+		cmd = None
 
-@export
-class Parser(metaclass=ExtendedType, slots=True):
-	_processor: "BaseProcessor"
+		tclProcedures = {"source"}
 
-	def __init__(self, processor: "BaseProcessor"):
-		self._processor = processor
+		# wait for first line
+		line = yield
+		# process preable
+		line = yield from self._preamble.Generator(line)
 
-	@readonly
-	def Processor(self) -> "BaseProcessor":
-		return self._processor
+		while True:
+			while True:
+				if isinstance(line, VivadoTclCommand):
+					if line._command == "synth_design":
+						self._commands[SynthesizeDesign] = (cmd := SynthesizeDesign(self))
+						line = yield next(gen := cmd.SectionDetector(line))
+						break
 
+				firstWord = line.Partition(" ")[0]
+				if firstWord in tclProcedures:
+					line = TclCommand.FromLine(line)
 
-@export
-class Preamble(Parser):
-	_toolVersion:   Nullable[YearReleaseVersion]
-	_startDatetime: Nullable[datetime]
+				line = yield line
 
-	_VERSION:   ClassVar[Pattern] = re_compile(r"""# Vivado v(\d+\.\d(\.\d)?) \(64-bit\)""")
-	_STARTTIME: ClassVar[Pattern] = re_compile(r"""# Start of session at: (\w+ \w+ \d+ \d+:\d+:\d+ \d+)""")
+			while True:
+				if line.StartsWith("synth_design:"):
+					lastLine = gen.send(line)
+					if LineKind.Last in line._kind:
+						line._kind ^= LineKind.Last
+					line = yield lastLine
+					break
 
-	def __init__(self, processor: "BaseProcessor"):
-		super().__init__(processor)
-
-		self._toolVersion =   None
-		self._startDatetime = None
-
-	@readonly
-	def ToolVersion(self) -> YearReleaseVersion:
-		return self._toolVersion
-
-	@readonly
-	def StartDatetime(self) -> datetime:
-		return self._startDatetime
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		rawMessage = line._message
-		if rawMessage.startswith("#----"):
-			line._kind = LineKind.SectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		line = yield line
-
-		while line is not None:
-			rawMessage = line._message
-
-			if (match := self._VERSION.match(rawMessage)) is not None:
-				self._toolVersion = YearReleaseVersion.Parse(match[1])
-				line._kind = LineKind.Normal
-			elif (match := self._STARTTIME.match(rawMessage)) is not None:
-				self._startDatetime = datetime.strptime(match[1], "%a %b %d %H:%M:%S %Y")
-				line._kind = LineKind.Normal
-			elif rawMessage.startswith("#----"):
-				line._kind = LineKind.SectionDelimiter | LineKind.Last
-				break
-			else:
-				line._kind = LineKind.Verbose
-
-			line = yield line
-
-		check = yield line
+				line = yield gen.send(line)
 
 
 @export
-class BaseDocument(BaseProcessor):
+class Document(Processor):
 	_logfile: Path
 	_lines:   List[Line]
-	# _duration: float
 
 	def __init__(self, logfile: Path) -> None:
 		super().__init__()
@@ -630,15 +254,10 @@ class BaseDocument(BaseProcessor):
 			with self._logfile.open("r", encoding="utf-8") as f:
 				content = f.read()
 
-			lines = content.splitlines()
-			next(generator := self.LineClassification(self.DocumentSlicer()))
-			self._lines = [generator.send(rawLine) for rawLine in lines]
+			self._lines = []
+			next(generator := self.LineClassification())
+			for rawLine in content.splitlines():
+				line = generator.send(rawLine)
+				self._lines.append(line)
 
 		self._duration = sw.Duration
-
-	def DocumentSlicer(self, line: Line) -> Generator[Union[Line, ProcessorException], Line, Line]:
-		while line is not None:
-			if line._kind is LineKind.Unprocessed:
-				line._kind = LineKind.Normal
-
-			line = yield line
